+++ analyzers/src/SonarAnalyzer.Utilities/Rules.Description/S5939_vb.html
