+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/UseArrayEmptyTest.cs
@@ -19,9 +19,9 @@
  */
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SonarAnalyzer.UnitTest.TestFramework;
 using CS = SonarAnalyzer.Rules.CSharp;
 using VB = SonarAnalyzer.Rules.VisualBasic;
+using SonarAnalyzer.UnitTest.TestFramework;
 
 namespace SonarAnalyzer.UnitTest.Rules
 {
@@ -30,21 +30,14 @@ namespace SonarAnalyzer.UnitTest.Rules
     {
         [TestMethod]
         [TestCategory("Rule")]
-        public void UseArrayEmptys_VB()
-        {
-            Verifier.VerifyAnalyzer(
-                @"TestCases\UseArrayEmpty.vb",
-                new VB.UseArrayEmpty());
-        }
+        public void UseArrayEmpty() =>
+            Verifier.VerifyAnalyzer(@"TestCases\UseArrayEmpty.cs", new CS.UseArrayEmpty());
 
         [TestMethod]
         [TestCategory("Rule")]
-        public void UseArrayEmptys_CS()
-        {
-            Verifier.VerifyAnalyzer(
-                @"TestCases\UseArrayEmpty.cs",
-                new CS.UseArrayEmpty(),
-                ParseOptionsHelper.FromCSharp8);
-        }
+        public void UseArrayEmpty_VB() =>
+            Verifier.VerifyAnalyzer(@"TestCases\UseArrayEmpty.vb", new VB.UseArrayEmpty());
     }
 }
+
+
